[API] add HoldResource() interface (#134)
Make resource holding an separate API. So we don't need to put
`resources...` to every `Init()` overload. This will cause big trouble
in the future when there are more and more `Init()` overloads. So I want
to change it before too many people use it.

This is a breaking change. Users should move resources from `Init()`
function to `ec_actor::HoldResource()`. See changes in
`test/scheduler_test.cc` for example.